Conflicts:
	src/kaapi.h
	src/kaapi_impl.h
	src/machine/mt/kaapi_mt_setcontext.c
	src/task/kaapi_stack_clear.c
	src/task/kaapi_stack_init.c

+++ src/misc/kaapi_hashvalue.c
@@ -124,6 +124,6 @@ kaapi_uint32_t kaapi_hash_value(const char * data)
 {
   if (data == 0) return 0;
 
-  int len = strlen( data );
+  int len = (int)strlen( data );
   return kaapi_hash_value_len( data, len );
 }
